third_party/WebKit/LayoutTests/fast/dom/dom-attribute-on-prototype-chain.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
var div = document.createElement("div");
shouldBeFalse('div.hasOwnProperty("id")');
shouldBeTrue('div.__proto__.__proto__.__proto__.hasOwnProperty("id")');

var desc = Object.getOwnPropertyDescriptor(Element.prototype, 'id');
shouldBeTrue('desc.get instanceof Function');
shouldBeTrue('desc.set instanceof Function');
shouldBeTrue('desc.enumerable');
shouldBeTrue('desc.configurable');

shouldThrow('xhr = new XMLHttpRequest(); xhr.__proto__ = HTMLDivElement.prototype; xhr.id');
shouldThrow('obj = Object.create(div); obj.id');

desc = Object.getOwnPropertyDescriptor(Element.prototype, 'id');
shouldThrow('desc.get.call({})');

var a = document.createElement('a');
shouldBeUndefined('desc.set.call(a, "abc")');
shouldBeEqualToString('desc.get.call(a)', 'abc');
</script>
</body>
</html>
